// habit-tracker-mobileapp/utils/enums/habitEnum.ts
export enum HabitCategory {
  General = "General",
  Health = "Health",
  Fitness = "Fitness",
  Study = "Study",
  Work = "Work",
  Finance = "Finance",
  SelfGrowth = "SelfGrowth",
  Social = "Social",
  Creative = "Creative",
  Environment = "Environment",
}

export enum HabitFrequency {
  Daily = "Daily",
  Weekly = "Weekly",
  Monthly = "Monthly",
}